chore: remove spring-jdbc dependencies
