+++ src/test/java/com/at/avro/integration/PgsqlIntegrationTest.java
@@ -5,12 +5,11 @@ import com.at.avro.DbSchemaExtractor;
 import com.at.avro.SchemaGenerator;
 import com.at.avro.config.AvroConfig;
 import org.flywaydb.core.Flyway;
+import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.testcontainers.containers.PostgreSQLContainer;
-
-import java.io.IOException;
 
 import static helper.Utils.classPathResourceContent;
 import static org.hamcrest.CoreMatchers.is;
@@ -21,9 +20,9 @@ public class PgsqlIntegrationTest {
     @ClassRule
     public static PostgreSQLContainer PGSQL = new PostgreSQLContainer();
 
-    final static private AvroConfig AVRO_CONFIG = new AvroConfig("pgsql");
+    private AvroConfig avroConfig = new AvroConfig("pgsql");
 
-    static private DbSchemaExtractor SCHEMA_EXTRACTOR;
+    private DbSchemaExtractor extractor;
 
     @BeforeClass
     public static void setupClass() {
@@ -31,19 +30,22 @@ public class PgsqlIntegrationTest {
             .dataSource(PGSQL.getJdbcUrl(), PGSQL.getUsername(), PGSQL.getPassword())
             .locations("classpath:pgsql/db/migration")
             .load().migrate();
+    }
 
-        SCHEMA_EXTRACTOR = new DbSchemaExtractor(PGSQL.getJdbcUrl(), PGSQL.getUsername(), PGSQL.getPassword());
+    @Before
+    public void setup() {
+        extractor = new DbSchemaExtractor(PGSQL.getJdbcUrl(), PGSQL.getUsername(), PGSQL.getPassword());
     }
 
     @Test
     public void testDefaultTabe() {
-        AvroSchema avroSchema = SCHEMA_EXTRACTOR.getForTable(AVRO_CONFIG, null, "default_table");
+        AvroSchema avroSchema = extractor.getForTable(avroConfig, null, "default_table");
         assertThat(SchemaGenerator.generate(avroSchema), is(classPathResourceContent("/pgsql/avro/default_table.avsc")));
     }
 
     @Test
     public void testArrayTable() {
-        AvroSchema avroSchema = SCHEMA_EXTRACTOR.getForTable(AVRO_CONFIG, null, "array_table");
+        AvroSchema avroSchema = extractor.getForTable(avroConfig, null, "array_table");
         assertThat(SchemaGenerator.generate(avroSchema), is(classPathResourceContent("/pgsql/avro/array_table.avsc")));
     }
 }
