+++ src/test/resources/hsql/avro/all_fields_default_null.avsc
