+++ src/test/resources/hsql/avro/all_nullable.avsc
