+++ src/test/resources/hsql/avro/camel_case.avsc
