+++ src/test/resources/hsql/avro/custom_formatter.avsc
