+++ src/test/resources/hsql/avro/custom_property.avsc
