+++ src/test/resources/hsql/avro/default.avsc
