+++ src/test/resources/hsql/avro/non_pretty_print.avsc
