+++ src/test/resources/hsql/avro/pretty_print_bigger_indent.avsc
